sequence gateway, new/open seq. impl.

+++ src/main/java/hu/boga/midiai/core/boundaries/SequenceDto.java
@@ -3,8 +3,8 @@ package hu.boga.midiai.core.boundaries;
 public class SequenceDto {
     public String name;
     public int resolution;
-    public int division;
-    public int tickLength;
+    public float division;
+    public long tickLength;
     public int ticksPerMeasure;
     public int ticksIn32nds;
 }
